[Client] Fixed giỏ hàng, thanh toán, notfoundPage, thanh toán vnpay, quên mật khẩu, gửi mail liên hệ

+++ Src/Controllers/Client/UserInfoController.php
@@ -8,6 +8,7 @@ use Src\Notifications\Notification;
 use Src\Models\Client\AddressModel;
 use Src\Validations\Client\DataValidation;
 use Src\Helpers\Client\AuthHelper;
+use Src\Models\Client\OrderModel;
 
 class UserInfoController extends BaseController
 {
@@ -18,12 +19,51 @@ class UserInfoController extends BaseController
     }
     public function orders()
     {
-        echo $this->view->render('Client/UserProfile/Orders', ['Name' => 'Bao']);
+        $userId = $_SESSION['user']['id'];
+        $orderModel = new OrderModel();
+        $data = $orderModel->getAllOrdersByUser($userId);
+        echo $this->view->render('Client/UserProfile/Orders', ['data' => $data]);
     }
     public function orderDetail()
     {
         echo $this->view->render('Client/UserProfile/OrderDetail', ['Name' => 'Bao']);
     }
+    public function cancelOrder()
+    {
+        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+            $orderId = $_POST['order_id'] ?? null;
+
+            if ($orderId) {
+                $orderModel = new OrderModel();
+                $userId = $_SESSION['user']['id'];
+
+                $order = $orderModel->getAllOrderByUserAndOrderId($orderId, $userId);
+
+                if (!empty($order)) {
+
+                    $orderDetails = $order[0];
+
+                    if ($orderDetails['order_status'] == 1) {
+                        $isCanceled = $orderModel->cancelOrder($orderId);
+
+                        if ($isCanceled) {
+                            Notification::success('Thành công', 'Đơn hàng đã được hủy.');
+                        } else {
+                            Notification::error('Thất bại', 'Lỗi khi hủy đơn hàng.');
+                        }
+                    } else {
+                        Notification::error('Thất bại', 'Không thể hủy đơn hàng này.');
+                    }
+                } else {
+                    Notification::error('Thất bại', 'Không tìm thấy đơn hàng của bạn.');
+                }
+            }
+        }
+
+        header('Location: /orders');
+        exit;
+    }
+
     public function address()
     {
         if (!isset($_SESSION['user'])) {
@@ -52,13 +92,13 @@ class UserInfoController extends BaseController
                 'district_id' => $district_id,
                 'ward_id' => $ward_id,
                 'user_id' => $_SESSION['user']['id'],
+                'address_username' => $_POST['address_username'],
                 'address' => $_POST['address'],
                 'phone' => $_POST['phone'],
                 'province_name' => $province_name,
                 'district_name' => $district_name,
                 'ward_name' => $ward_name
             ];
-
             $result = $AddressModel->createAddress($data);
             if ($result) {
                 Notification::success('Thành công', 'Đã thêm địa chỉ');
@@ -78,8 +118,26 @@ class UserInfoController extends BaseController
     public function deleteAddress($id)
     {
         try {
-            $id = (int)$id;
+            if (is_array($id)) {
+                $id = reset($id);
+            }
 
+            $id = (int) $id;
+
+            if ($id <= 0) {
+                throw new \Exception('ID không hợp lệ');
+            }
+            $orderModel = new OrderModel();
+            $relatedOrders = $orderModel->getOrdersByAddressId($id);
+    
+            if (!empty($relatedOrders)) {
+                Notification::error(
+                    'Thao tác thất bại',
+                    'Không thể xóa địa chỉ này vì đang được sử dụng trong các đơn hàng.'
+                );
+                header('Location: /address');
+                exit;
+            }
             $AddressModel = new AddressModel();
             $result = $AddressModel->delete($id);
 
@@ -89,12 +147,13 @@ class UserInfoController extends BaseController
                 Notification::error('Thất bại', 'Đã xảy ra lỗi khi xóa địa chỉ');
             }
         } catch (\Exception $e) {
-            Notification::error('Thất bại', 'Đã xảy ra lỗi không mong muốn');
+            Notification::error('Thất bại', 'Lỗi: ' . $e->getMessage());
         }
 
         header('location: /address');
         exit();
     }
+
     public function changePassword()
     {
         echo $this->view->render('Client/UserProfile/ChangePassword', ['Name' => 'Bao']);
@@ -115,11 +174,10 @@ class UserInfoController extends BaseController
             ];
 
             $result = AuthHelper::updatePassword($data);
-            
+
             if ($result) {
                 Notification::success('Thành công', 'Đã cập nhật mật khẩu');
             }
-            
         } catch (\Exception $e) {
             Notification::error('Thất bại', 'Đã xảy ra lỗi khi cập nhật mật khẩu');
         }
